refactoring view-templates
