+++ src/templates/trip-events-list-template.js
@@ -1,0 +1,3 @@
+const createEventListTemplate = () => ('<ul class="trip-events__list"></ul>');
+
+export { createEventListTemplate };
