+++ src/view/sort-view.js
@@ -1,33 +1,34 @@
 import { createElement } from '../render.js';
+import { createSortTemplate } from '../templates/sort-view-template.js';
 
-const createSortTemplate = () => (
-  `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
-    <div class="trip-sort__item  trip-sort__item--day">
-      <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked>
-      <label class="trip-sort__btn" for="sort-day">Day</label>
-    </div>
+// const createSortTemplate = () => (
+//   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
+//     <div class="trip-sort__item  trip-sort__item--day">
+//       <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked>
+//       <label class="trip-sort__btn" for="sort-day">Day</label>
+//     </div>
 
-    <div class="trip-sort__item  trip-sort__item--event">
-      <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" disabled>
-      <label class="trip-sort__btn" for="sort-event">Event</label>
-    </div>
+//     <div class="trip-sort__item  trip-sort__item--event">
+//       <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" disabled>
+//       <label class="trip-sort__btn" for="sort-event">Event</label>
+//     </div>
 
-    <div class="trip-sort__item  trip-sort__item--time">
-      <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time">
-      <label class="trip-sort__btn" for="sort-time">Time</label>
-    </div>
+//     <div class="trip-sort__item  trip-sort__item--time">
+//       <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time">
+//       <label class="trip-sort__btn" for="sort-time">Time</label>
+//     </div>
 
-    <div class="trip-sort__item  trip-sort__item--price">
-      <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price">
-      <label class="trip-sort__btn" for="sort-price">Price</label>
-    </div>
+//     <div class="trip-sort__item  trip-sort__item--price">
+//       <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price">
+//       <label class="trip-sort__btn" for="sort-price">Price</label>
+//     </div>
 
-    <div class="trip-sort__item  trip-sort__item--offer">
-      <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
-      <label class="trip-sort__btn" for="sort-offer">Offers</label>
-    </div>
-  </form>`
-);
+//     <div class="trip-sort__item  trip-sort__item--offer">
+//       <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
+//       <label class="trip-sort__btn" for="sort-offer">Offers</label>
+//     </div>
+//   </form>`
+// );
 
 export default class SortView {
