+++ src/view/trip-events-list.js
@@ -1,6 +1,7 @@
 import { createElement } from '../render.js';
+import { createEventListTemplate } from '../templates/trip-events-list-template.js';
 
-const createEventListTemplate = () => ('<ul class="trip-events__list"></ul>');
+// const createEventListTemplate = () => ('<ul class="trip-events__list"></ul>');
 
 export default class TripEventListView {
